+++ noodles-bgzf/src/io.rs
@@ -3,6 +3,7 @@
 mod buf_read;
 pub mod indexed_reader;
 mod multithreaded_reader;
+pub mod multithreaded_writer;
 mod read;
 pub mod reader;
 mod seek;
@@ -10,5 +11,6 @@ pub mod writer;
 
 pub use self::{
     buf_read::BufRead, indexed_reader::IndexedReader, multithreaded_reader::MultithreadedReader,
-    read::Read, reader::Reader, seek::Seek, writer::Writer,
+    multithreaded_writer::MultithreadedWriter, read::Read, reader::Reader, seek::Seek,
+    writer::Writer,
 };